refactor: optimize template dependencies and architectural structure
- package.json: Removed sass, lodash.merge, and @types/lodash.merge dependencies.
- src/app/metadata.ts: Removed 'Sass' from SEO keywords.
- src/app/page.tsx: Updated ThemeSwitch import path.
- src/assets/styles/globals.css: Removed custom hamburger menu CSS in favor of pure Tailwind.
- src/components/global/ThemeSwitch.tsx: Moved from src/components/features/ to comply with project structure.
- src/layout/navbar.tsx: Refactored desktop links to use Next.js Link and rewritten hamburger menu using pure Tailwind classes.
- src/layout/sidebar.tsx: Updated ThemeSwitch import path.
- src/providers/AppProviders.tsx: Replaced lodash.merge with native object expansion for RainbowKit theme customization.
