+++ package.json
@@ -29,7 +29,6 @@
     "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.18.0",
     "@tailwindcss/postcss": "^4.2.1",
-    "@types/lodash.merge": "^4.6.9",
     "@types/node": "^25.3.3",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
@@ -37,11 +36,9 @@
     "eslint-config-next": "^15.1.7",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-lingui": "^0.11.0",
-    "lodash.merge": "^4.6.2",
     "postcss": "^8.5.8",
     "prettier": "^3.8.1",
     "prettier-plugin-tailwindcss": "^0.7.2",
-    "sass": "^1.97.3",
     "tailwindcss": "^4.2.1",
     "typescript": "^5.9.3"
   },
