+++ src/assets/styles/globals.css
@@ -50,28 +50,3 @@ body {
   margin: 0;
   padding: 0;
 }
-
-/* Hamburger menu styles */
-.line {
-  width: 100%;
-  height: 3px;
-  background-color: #333;
-  border-radius: 12px;
-  transition: all 0.3s ease-in-out;
-}
-
-.hamburger.open .line:nth-of-type(1) {
-  transform: rotate(45deg) translate(7px, 5px);
-}
-
-.hamburger.open .line:nth-of-type(2) {
-  opacity: 0;
-}
-
-.hamburger.open .line:nth-of-type(3) {
-  transform: rotate(-45deg) translate(7px, -5px);
-}
-
-.dark .line {
-  background-color: #fff;
-}
