+++ src/providers/AppProviders.tsx
@@ -3,13 +3,15 @@
 import { type JSX } from 'react';
 import { RainbowKitProvider, lightTheme, Theme } from '@rainbow-me/rainbowkit';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import merge from 'lodash.merge';
 import { ThemeProvider } from 'next-themes';
 import { WagmiProvider } from 'wagmi';
 import { wagmiConfig } from '@/lib/constants/wagmiConfig';
 
-const theme = merge(lightTheme(), {
+const baseTheme = lightTheme();
+const theme: Theme = {
+  ...baseTheme,
   colors: {
+    ...baseTheme.colors,
     accentColor: '#FF801F',
     accentColorForeground: '#fff',
     actionButtonSecondaryBackground: '#DADDD8',
@@ -19,7 +21,7 @@ const theme = merge(lightTheme(), {
     connectButtonText: '#000',
     connectButtonTextError: '#FF494A',
   },
-} as Theme);
+};
 
 const queryClient = new QueryClient();
